gitignore: be more precise with the rules
It is bad to have .* as a blanket ignore. For example, whenever you try
to add anything starting with dot (e.g., .travis.yml), you get a warning
which you must explicitly override.

Anyway, it is better to be specific about what sorts of files should be
ignore. That way, git does not accidentally ignore something important.