Improve reporting of module details
We want to know what the last values were, as well as what is
still unresolved at the time a dialog is supposed to pop up.

This approach adds classes WebCommandInfo and WebCommandModuleItem,
enabling the Jackson serializer to access the information on inputs,
e.g. to find out whether a specific input has been resolved or not.

Co-authored-by: Petr Bainar <petr.bainar@hotmail.com>