ServerArgument: wait for headless server to stop
We accomplish this by registering a startup operation that blocks
until the server shuts down, but only while in headless mode.

In this way, when "--server" is passed together with "--headless",
the application with continue running until the server terminates.
And when "--server" is passed without "--headless", the server
starts up, but does not block the application from shutting down.