Remove package-lock.json, to reduce CVE alerts
I'm fully aware that this does not actually address security
vulnerabilities. But imagej-server is not being actively maintained
right now. And people are free to use the package.json to let the solver
try to install dependencies at the most currently recommended versions.
My guess though is that there are 0 users of the webNew client, sadly.